requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-import-export==3.0.1
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
